+++ src/test/java/nextstep/subway/acceptance/SectionAcceptanceTest.java
@@ -1,5 +1,10 @@
 package nextstep.subway.acceptance;
 
+import static nextstep.subway.utils.LineAcceptanceHelper.노선_생성;
+import static nextstep.subway.utils.LineAcceptanceHelper.노선_조회;
+import static nextstep.subway.utils.SectionAcceptanceHelper.구간_제거;
+import static nextstep.subway.utils.SectionAcceptanceHelper.구간_추가;
+import static nextstep.subway.utils.StationAcceptanceTest.지하철역_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
